src/entities/Counter/ui/Counter.tsx
import React, { FC } from "react";
import { useTranslation } from "react-i18next";
import { useDispatch, useSelector } from "react-redux";
import { Button } from "shared/ui/Button/Button";
import { counterActions } from "../model/slice/counterSlice";
import { getCounterValue } from "../model/selectors/getCounterValue/getCounterValue";

export const Counter: FC = () => {
  const { t } = useTranslation();
  const dispatch = useDispatch();
  const counterValue = useSelector(getCounterValue);

  const increment = () => {
    dispatch(counterActions.increment());
  };

  const decrement = () => {
    dispatch(counterActions.decrement());
  };
  return (
    <div>
      <h1 data-testid="counter-value">{counterValue}</h1>
      <Button onClick={increment} data-testid="increment">
        {t("инкремент")}
      </Button>
      <Button onClick={decrement} data-testid="decrement">
        {t("декремент")}
      </Button>
    </div>
  );
};
